minor clean up, refactor, delete old classes. not functionally complete, missing tags
