+++ app/src/main/java/com/example/kit/SelectListener.java
@@ -1,9 +1,6 @@
 package com.example.kit;
 
-import android.view.View;
-
 import com.example.kit.data.Item;
-import com.example.kit.database.ItemViewHolder;
 
 
 /** SelectListener is an interface that is used to implement UI interactions
