+++ doc/javaDocs/com/example/kit/database/FirestoreManager.html
@@ -72,7 +72,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class FirestoreManager" class="title">Class FirestoreManager</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.example.kit.database.FirestoreManager</div>
+<div class="inheritance">com.example.kit.data.FirestoreManager</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
