+++ doc/javaDocs/com/example/kit/database/ItemFirestoreAdapter.html
@@ -126,7 +126,7 @@ loadScripts(document, 'script');</script>
 <div class="table-header col-second">Method</div>
 <div class="table-header col-last">Description</div>
 <div class="col-first even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>protected void</code></div>
-<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#onBindViewHolder(com.example.kit.database.ItemViewHolder,int,com.example.kit.data.Item)" class="member-name-link">onBindViewHolder</a><wbr>(<a href="ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a>&nbsp;holder,
+<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#onBindViewHolder(com.example.kit.ItemViewHolder,int,com.example.kit.data.Item)" class="member-name-link">onBindViewHolder</a><wbr>(<a href="ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a>&nbsp;holder,
  int&nbsp;position,
  <a href="../data/Item.html" title="class in com.example.kit.data">Item</a>&nbsp;model)</code></div>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
@@ -187,7 +187,7 @@ loadScripts(document, 'script');</script>
 <h2>Method Details</h2>
 <ul class="member-list">
 <li>
-<section class="detail" id="onBindViewHolder(com.example.kit.database.ItemViewHolder,int,com.example.kit.data.Item)">
+<section class="detail" id="onBindViewHolder(com.example.kit.ItemViewHolder,int,com.example.kit.data.Item)">
 <h3>onBindViewHolder</h3>
 <div class="member-signature"><span class="modifiers">protected</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">onBindViewHolder</span><wbr><span class="parameters">(@NonNull
  <a href="ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a>&nbsp;holder,
