+++ doc/javaDocs/com/example/kit/database/ItemViewHolder.html
@@ -73,7 +73,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">androidx.recyclerview.widget.RecyclerView.ViewHolder
-<div class="inheritance">com.example.kit.database.ItemViewHolder</div>
+<div class="inheritance">com.example.kit.ItemViewHolder</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
@@ -131,7 +131,7 @@ loadScripts(document, 'script');</script>
  outside the view holder</div>
 </div>
 <div class="col-first even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code>void</code></div>
-<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#setupListeners(com.example.kit.SelectListener,com.example.kit.database.ItemViewHolder,com.example.kit.data.Item)" class="member-name-link">setupListeners</a><wbr>(<a href="../SelectListener.html" title="interface in com.example.kit">SelectListener</a>&nbsp;listener,
+<div class="col-second even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4"><code><a href="#setupListeners(com.example.kit.SelectListener,com.example.kit.ItemViewHolder,com.example.kit.data.Item)" class="member-name-link">setupListeners</a><wbr>(<a href="../SelectListener.html" title="interface in com.example.kit">SelectListener</a>&nbsp;listener,
  <a href="ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a>&nbsp;holder,
  <a href="../data/Item.html" title="class in com.example.kit.data">Item</a>&nbsp;model)</code></div>
 <div class="col-last even-row-color method-summary-table method-summary-table-tab2 method-summary-table-tab4">
@@ -191,7 +191,7 @@ loadScripts(document, 'script');</script>
 </section>
 </li>
 <li>
-<section class="detail" id="setupListeners(com.example.kit.SelectListener,com.example.kit.database.ItemViewHolder,com.example.kit.data.Item)">
+<section class="detail" id="setupListeners(com.example.kit.SelectListener,com.example.kit.ItemViewHolder,com.example.kit.data.Item)">
 <h3>setupListeners</h3>
 <div class="member-signature"><span class="modifiers">public</span>&nbsp;<span class="return-type">void</span>&nbsp;<span class="element-name">setupListeners</span><wbr><span class="parameters">(<a href="../SelectListener.html" title="interface in com.example.kit">SelectListener</a>&nbsp;listener,
  <a href="ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a>&nbsp;holder,
