+++ doc/javaDocs/index-files/index-12.html
@@ -86,7 +86,7 @@ loadScripts(document, 'script');</script>
 <dd>&nbsp;</dd>
 <dt><a href="../com/example/kit/data/Item.html#setTags(java.util.ArrayList)" class="member-name-link">setTags(ArrayList&lt;String&gt;)</a> - Method in class com.example.kit.data.<a href="../com/example/kit/data/Item.html" title="class in com.example.kit.data">Item</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/example/kit/database/ItemViewHolder.html#setupListeners(com.example.kit.SelectListener,com.example.kit.database.ItemViewHolder,com.example.kit.data.Item)" class="member-name-link">setupListeners(SelectListener, ItemViewHolder, Item)</a> - Method in class com.example.kit.database.<a href="../com/example/kit/database/ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a></dt>
+<dt><a href="../com/example/kit/database/ItemViewHolder.html#setupListeners(com.example.kit.SelectListener,com.example.kit.ItemViewHolder,com.example.kit.data.Item)" class="member-name-link">setupListeners(SelectListener, ItemViewHolder, Item)</a> - Method in class com.example.kit.database.<a href="../com/example/kit/database/ItemViewHolder.html" title="class in com.example.kit.database">ItemViewHolder</a></dt>
 <dd>
 <div class="block">This sets up listeners for individual UI elements for each item,
  this requires a <a href="../com/example/kit/SelectListener.html" title="interface in com.example.kit"><code>SelectListener</code></a></div>
