+++ doc/javaDocs/index-files/index-9.html
@@ -59,7 +59,7 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><a href="../com/example/kit/SelectListener.html#onAddTagClick()" class="member-name-link">onAddTagClick()</a> - Method in interface com.example.kit.<a href="../com/example/kit/SelectListener.html" title="interface in com.example.kit">SelectListener</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/example/kit/database/ItemFirestoreAdapter.html#onBindViewHolder(com.example.kit.database.ItemViewHolder,int,com.example.kit.data.Item)" class="member-name-link">onBindViewHolder(ItemViewHolder, int, Item)</a> - Method in class com.example.kit.database.<a href="../com/example/kit/database/ItemFirestoreAdapter.html" title="class in com.example.kit.database">ItemFirestoreAdapter</a></dt>
+<dt><a href="../com/example/kit/database/ItemFirestoreAdapter.html#onBindViewHolder(com.example.kit.ItemViewHolder,int,com.example.kit.data.Item)" class="member-name-link">onBindViewHolder(ItemViewHolder, int, Item)</a> - Method in class com.example.kit.database.<a href="../com/example/kit/database/ItemFirestoreAdapter.html" title="class in com.example.kit.database">ItemFirestoreAdapter</a></dt>
 <dd>
 <div class="block">Populates the <a href="../com/example/kit/database/ItemViewHolder.html" title="class in com.example.kit.database"><code>ItemViewHolder</code></a> with the <a href="../com/example/kit/data/Item.html" title="class in com.example.kit.data"><code>Item</code></a></div>
 </dd>
